refactor: Move dynamic symbol creation to elf.rs
